dividing the procedure in modules
the entire procedure has been divided in singular modules, everyone of which is responsible of a particular function inside the context of sensors detection via Raspberry and Arduino.
A functional backup folder as been created for saving a copy of the old version of the script in which all functions were contained.